refactor: split controller and message chain
